backend/requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Database
mysqlclient==2.2.0
psycopg2-binary==2.9.9

# File processing
Pillow==10.1.0
PyPDF2==3.0.1
yt-dlp==2024.8.6

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Azure specific
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Performance and caching
redis==5.0.1
celery==5.3.4

# Monitoring and logging
azure-monitor-opentelemetry==1.2.0
applicationinsights==0.11.10

# Security
cryptography==41.0.7
